# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/lembretes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lembretes"
                ],
                "summary": "listlembretes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insere os lembretes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lembretes"
                ],
                "summary": "insertlembretes",
                "parameters": [
                    {
                        "description": "struct",
                        "name": "Lembretes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Lembrete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lembretes/{id}": {
            "put": {
                "description": "Altera os lembretes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lembretes"
                ],
                "summary": "updatelembretes",
                "parameters": [
                    {
                        "description": "struct",
                        "name": "Lembretes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Lembrete"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Lembrete ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Exclui os lembretes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lembretes"
                ],
                "summary": "deletelembretes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lembrete ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/usuario": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "listusuario",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insere os usuarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "insertusuario",
                "parameters": [
                    {
                        "description": "struct",
                        "name": "Usuarios",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Usuario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/usuario/{id}": {
            "put": {
                "description": "Altera os usuarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "updateusuario",
                "parameters": [
                    {
                        "description": "struct",
                        "name": "Usuarios",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Usuario"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Usuario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta os usuarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "deleteusuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Usuario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Lembrete": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "usuarioID": {
                    "type": "integer"
                }
            }
        },
        "main.Usuario": {
            "type": "object",
            "properties": {
                "lembretes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Lembrete"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "sobrenome": {
                    "type": "string"
                }
            }
        }
    }
}